#250 refactor so correct python path is sent
