+++ src/PreviewManager.ts
@@ -1,13 +1,12 @@
 "use strict"
 import {PythonEvaluator} from "arepl-backend"
-import {EOL} from "os"
-import { isAbsolute, sep } from "path";
 import * as vscode from "vscode"
-import { PreviewContainer } from "./previewContainer";
+import { PreviewContainer } from "./previewContainer"
 import Reporter from "./telemetry"
 import {ToAREPLLogic} from "./toAREPLLogic"
-import vscodeUtils from "./vscodeUtilities";
-import { PythonShell } from "python-shell";
+import vscodeUtils from "./vscodeUtilities"
+import areplUtils from "./areplUtilities"
+import { PythonShell } from "python-shell"
 import {settings} from "./settings"
 
 /**
@@ -62,7 +61,7 @@ export default class PreviewManager {
         this.startAndBindPython()
 
         if(this.pythonEditorDoc.isUntitled && this.pythonEditorDoc.getText() == "") {
-            await this.insertDefaultImports(this.pythonEditor)
+            await areplUtils.insertDefaultImports(this.pythonEditor)
             // waiting for this to complete so i dont accidentily trigger
             // the edit doc handler when i insert imports
         }
@@ -133,41 +132,11 @@ export default class PreviewManager {
         }
     }
 
-    getPythonPath(){
-        let pythonPath = settings().get<string>("pythonPath")
-
-        const pythonExtSettings = vscode.workspace.getConfiguration("python", null);
-        const pythonExtPythonPath = pythonExtSettings.get<string>('pythonPath')
-        if(pythonExtPythonPath && !pythonPath) pythonPath = pythonExtPythonPath
-
-        if(pythonPath){
-            pythonPath = pythonPath.replace("${workspaceFolder}", vscodeUtils.getCurrentWorkspaceFolder())
-
-            let envVar = pythonPath.match(/\${env:([^}]+)}/)
-            if(envVar){
-                pythonPath = pythonPath.replace(envVar[1], process.env[envVar[1]])
-            }
-
-            // not needed anymore but here for backwards compatability. Remove in 2020
-            pythonPath = pythonPath.replace("${python.pythonPath}", pythonExtSettings.get('pythonPath'))
-
-            // if its a relative path, make it absolute
-            if(pythonPath.includes(sep) && !isAbsolute(pythonPath)){
-                pythonPath = vscodeUtils.getCurrentWorkspaceFolder() + sep + pythonPath
-            }
-        }
-        else{
-            pythonPath = PythonShell.defaultPythonPath
-        }
-
-        return pythonPath
-    }
-
     /**
      * starts AREPL python backend and binds print&result output to the handlers
      */
     private startAndBindPython(){
-        const pythonPath = this.getPythonPath()
+        const pythonPath = areplUtils.getPythonPath()
         const pythonOptions = settings().get<string[]>("pythonOptions")
 
         PythonShell.getVersion(`"${pythonPath}"`).then((out)=>{
@@ -264,28 +233,6 @@ export default class PreviewManager {
         }, this, this.subscriptions)
     }
 
-    private insertDefaultImports(editor: vscode.TextEditor){
-        return editor.edit((editBuilder) => {
-            let imports = settings().get<string[]>("defaultImports")
-
-            imports = imports.filter(i => i.trim() != "")
-            if(imports.length == 0) return
-
-            imports = imports.map(i => {
-                const words = i.split(" ")
-
-                // python import syntax: "import library" or "from library import method"
-                // so if user didnt specify import we will do that for them :)
-                if(words[0] != "import" && words[0] != "from" && words[0].length > 0){
-                    i = "import " + i
-                }
-
-                return i
-            })
-
-            editBuilder.insert(new vscode.Position(0, 0), imports.join(EOL) + EOL)
-        })
-    }
 
     private onAnyDocChange(event: vscode.TextDocument){
         if(event == this.pythonEditorDoc){
